requirements.txt
# Core face recognition libraries
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1

# Modern alternatives
mediapipe==0.10.7
insightface==0.7.3
onnxruntime==1.16.3

# Web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Database
sqlite3  # Built-in with Python
sqlalchemy==2.0.23

# Security and utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0

# Image processing
scikit-image==0.21.0
matplotlib==3.7.2

# Development and testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0
